requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7

# Data Validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Environment Variables
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Type Checking
mypy==1.7.0

# Code Formatting
black==23.11.0
isort==5.12.0

# Linting
flake8==6.1.0